refactor: remove redundant code and add constant for failed request assertion
